refactor: #890 — split LiveTutor into useResizablePanel, useLiveTutorSpeech, useTtsPlayback hooks (#895)
Extract three custom hooks and two utility modules from LiveTutor.tsx (1589 lines → 1075 lines):

- useResizablePanel.ts (80 lines) — drag-to-resize panel logic; also applied to FullReading.tsx to eliminate duplication
- useLiveTutorSpeech.ts (313 lines) — continuous STT session with auto-reconnect, per-sentence isFinal eval, real-time LCS overlay
- useTtsPlayback.ts (172 lines) — Cloud TTS + Web Speech API fallback with rAF cursor animation
- utils/liveTutorPools.ts (46 lines) — canned feedback response pools
- utils/liveTutorHelpers.ts (128 lines) — buildRealtimeOverlay, extractPracticeChars, calcSpeakingProgress, IS_PUNCT

Zero behavior changes. All evaluation logic, localStorage persistence, and JSX remain in LiveTutor.tsx.
